# hungry-wolf-root/hungry-wolf-facts/swagger/hungry-wolf-feed-api/hungry-wolf-feed-api.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: Hungry Wolf feed API
  description: Hungry Wolf feed API
  termsOfService: https://www.hungryWolf.com/terms-conditions/
host: api.feed.hungrywolf.com
basePath: /
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  APISecurity:
    type: basic
security:
  - APISecurity: []
paths:
...
  /customer/cartToOrder/:
    post:
      tags:
        - Order 
      description: Create Order
      operationId: createOrder
      produces:
        - application/json
      parameters:
         - name: request
           in: body
           description: Process Details
           required: true
           schema:
            $ref: '#/definitions/Order'
      responses:
        '200':
          description: OK
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/errorModel'   
...
definitions:
...
  Order:
    type: object
    properties:
      id:
        type: integer
      orderHwrn:
        type: string
        example: "0012459-1235689-154263"
      cartId:
        type: integer
      wolfId:
        type: integer
      createdDate:
        type: string
        format: date-time
      orderPlacedDate:
        type: string
        format: date-time
      modifiedDate:  
        type: string
        format: date-time
      status:
        type: string
        enum:
          - O
          - C
          - D
          - I
          - P
        description: 
          O - OPEN
          C - CANCELLED
          D - DELETED
          I - INPROGRESS
          P - PARTIALLY DELIVERED
      applyDiscount:
        type: string
        example: "HUNGRYCHALLENGE"
        enum:
          - NN
          - FF
          - HG
          - BG
          - HD
          - AH
          - ME
        description:
          NN - NONE
          FF - FIRSTFIV
          HG - HUNGRYCHALLENGE
          BG - BETHECHANGE
          HD - HAPPYDA
          AH - AWAYFROMHOME
          ME - MONTHEND
      items:
        type: array
        items:
          $ref: '#/definitions/OrderItems'
  OrderItems:
    type: object
    properties:
      cartItemId:
        type: integer
        example: "123456"
      mealDetails:
        $ref: '#/definitions/MealDetails'
      isCustomChanges:
        type: boolean
      chefId:
        type: integer
      quantity:
        type: integer
      createdDate:
        type: string
        format: date-time
      createdUser:
        type: string
      modifiedDate:
        type: string
        format: date-time
      modifiedUser:
        type: string  
      scheduleDeliveryDate:
        type: string
        format: date
        example: "31/03/2019"
      actualDelivertDate:
        type: string
        example: "31/03/2019"
      deliveryMethod:
        type: string
        enum:
          - D
          - P
        description:
          D - delivery
          P - Pickup
...
  MealDetails:
    type: object
    properties:
      meal:
        $ref: '#/definitions/Meal'
      mealSummary:    
        $ref: '#/definitions/MealSummary'
      mealServing:
        $ref: '#/definitions/MealServing'
      mealDelivery:
        $ref: '#/definitions/MealDelivery'
      pricing:
        $ref: '#/definitions/PricingInfo'  
...
  PricingInfo:
    type: object
    properties:
      id:
        type: integer
      rateCd:
        type: string
        format: string
      rate:
        type: string
        format: string
      currency:
        type: string
        format: string
      rateType:
        type: string
        format: string
  Meal:
    type: object
    properties:
      id:
        type: string
      title:
        type: string
        description: asdfasd
        example: adadas
      image:
        type: string
        description: url for the meal
        example: 001
      mealDate:
        type: string
        format: date-time
        description: Meal Start Date
        example: 2018-06-22T00:00:00.000Z  
      chef:
        $ref: '#/definitions/Chef'
      tags:
        $ref: '#/definitions/MealTags'
...
  MealTags:
    type: object
    properties:
      id:
        type: integer
      mealId:
        type: integer  
      tags:
        type: array
        items:
          type: string
          format: string
...
  Chef:    
    type: object
    properties:
      id:
        type: integer
      userId:
        type: string
      avatar:
        type: string
      rating:
        $ref: '#/definitions/ChefRating'
  ChefRating:
    type: object
    properties:
      id:
        type: integer
      overall:
        type: number
        format: int32
      detail:
        $ref: '#/definitions/ChefRatingDetails'  
  ChefRatingDetails:
    type: object
    properties:
      id:
        type: integer
      mealId:
        type: integer
      chefId:
        type: integer
      items:
        type: array
        items:
          $ref: '#/definitions/ChefRatingDetail'
  ChefRatingDetail:
    type: object
    properties:
      id:
        type: integer
      mealId:
        type: integer
      rating:
        type: number
        format: int32
        description: amount 
        example: 15.25 
      ratingFactorDesc:
        type: string
        example: "Quantity" 
      ratingCd:
        type: string
        example: "QTY" 
  MealSummary:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: '#/definitions/MealItems' 
  MealServing:
    type: object
    properties:
      id:
        type: integer
      mealId:
        type: integer  
      available:
        type: number
        format: int32
        description: amount 
        example: 15.25
      timeLeft:
        type: string
        format: string
        description: time in mins 
        example: 15.25
      mealDate:
        type: string
        format: date-time
        description: Meal Order deadline Date
        example: 2018-06-22T00:00:00.000Z 
      orderClosedBy:
        type: string
        format: date-time
        description: Meal Order deadline Date
        example: 2018-06-22T00:00:00.000Z
  MealDelivery:
    type: object
    properties:
      id:
        type: integer
      mealId:
        type: integer  
      deliveryType:
        type: string
        format: string
        description: delivery type
        example: pickup, delivery 
      mealDate:
        type: string
        format: date-time
        description: Meal Order deadline Date
        example: 2018-06-22T00:00:00.000Z       
  MealItems:
    type: object
    properties:
      mealId:
        type: integer
      itemId:
        type: string
        description:  sdfsdf
      desc:
        type: string
        description: meal item desc
      dishType:
        type: string
        description: gravy/sauce
      measurement:
        type: string
        description: serving measurement
  errorModel:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        example: 500
      message:
        type: string
        example: Illegal Argument Exception